+++ lib/bugsnag.rb
@@ -44,7 +44,13 @@ module Bugsnag
     # Notify of an exception unless it should be ignored
     def notify_or_ignore(exception, overrides=nil, request_data=nil)
       notification = Notification.new(exception, configuration, overrides, request_data)
-      notification.deliver unless notification.ignore?
+
+      unless notification.ignore?
+        notification.deliver
+        notification
+      else
+        false
+      end
     end
 
     # Auto notify of an exception, called from rails and rack exception